Fix for #197 Print Menu file list, when there were exactely 3 files or folders on the sd card. This was caused by a misinterpreted comparision between uint16_t and integer.

+++ Marlin/src/lcd/anycubic_touchscreen.h
@@ -281,7 +281,7 @@ private:
   int serial3_count = 0;
   char *TFTstrchr_pointer;
   char FlagResumFromOutage = 0;
-  uint16_t filenumber = 0;
+  int filenumber = 0;
   unsigned long starttime = 0;
   unsigned long stoptime = 0;
   uint8_t tmp_extruder = 0;
